refactor(app-config): delete package, inline constants into consumers
PINECONE_CONFIG → app-pinecone, EMBEDDING_CONFIG → app-embed.
The YAML parser (glob, parse, schema) and remaining configs
(CHUNKING, GITHUB, RERANK) had no external consumers.

Entire-Checkpoint: 8a8810ae8113
