+++ .changeset/pre.json
@@ -15,7 +15,6 @@
     "@repo/app-ai": "0.1.0",
     "@repo/app-ai-types": "0.1.0",
     "@repo/app-api-key": "0.1.0",
-    "@repo/app-config": "0.1.0",
     "@repo/app-embed": "0.1.0",
     "@repo/app-octokit-github": "0.1.0",
     "@repo/app-pinecone": "0.1.0",
