--- packages/app-config/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@repo/app-config",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": {
-      "types": "./src/index.ts",
-      "default": "./src/index.ts"
-    },
-    "./parse": "./src/parse.ts",
-    "./schema": "./src/schema.ts"
-  },
-  "license": "FSL-1.1-Apache-2.0",
-  "scripts": {
-    "clean": "git clean -xdf .cache .turbo node_modules",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@repo/app-validation": "workspace:*",
-    "@vendor/observability": "workspace:*",
-    "fast-glob": "^3.3.2",
-    "neverthrow": "catalog:",
-    "yaml": "^2.8.3",
-    "zod": "catalog:"
-  },
-  "devDependencies": {
-    "@repo/typescript-config": "workspace:*",
-    "@types/node": "catalog:",
-    "typescript": "catalog:"
-  }
-}
