--- packages/app-config/src/schema.ts
@@ -1,65 +1,0 @@
-/**
- * Zod schemas for lightfast.yml validation
- *
- * @see docs/architecture/phase1/dx-configuration.md
- */
-
-import { z } from "zod";
-
-/**
- * Zod schema for lightfast.yml configuration
- *
- * @example
- * ```yaml
- * version: 1
- * store: docs-site
- * include:
- *   - apps/docs/src/content/docs/**\/*.mdx
- *   - apps/docs/src/content/api/**\/*.mdx
- * ```
- */
-export const LightfastConfigSchema = z.object({
-  /**
-   * Configuration version (currently only version 1 is supported)
-   */
-  version: z.literal(1),
-
-  /**
-   * Human-readable store name (unique per organization)
-   * Used as the store key in the system
-   *
-   * Constraints:
-   * - Max 20 characters
-   * - Only lowercase alphanumeric and hyphens
-   * - No leading/trailing hyphens
-   * - No consecutive hyphens
-   *
-   * @example "docs-site" or "api-reference"
-   */
-  store: z
-    .string()
-    .min(1, "Store name must not be empty")
-    .max(20, "Store name must be 20 characters or less")
-    .regex(
-      /^[a-z0-9-]+$/,
-      "Store name must be lowercase alphanumeric with hyphens only"
-    )
-    .refine(
-      (name) => !/^-|-$|--/.test(name),
-      "Store name cannot have leading/trailing/consecutive hyphens"
-    ),
-
-  /**
-   * Array of glob patterns (repo-relative) for files to include in ingestion
-   * Must contain at least one pattern
-   * @example ["apps/docs/src/content/**\/*.mdx"]
-   */
-  include: z
-    .array(z.string().min(1, "Glob pattern must not be empty"))
-    .min(1, "At least one include pattern is required"),
-});
-
-/**
- * TypeScript type inferred from LightfastConfigSchema
- */
-export type LightfastConfig = z.infer<typeof LightfastConfigSchema>;
