--- packages/app-config/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "@repo/typescript-config/base.json",
-  "include": ["src"],
-  "exclude": ["node_modules"]
-}
