--- packages/app-config/turbo.json
@@ -1,5 +1,0 @@
-{
-  "extends": ["//"],
-  "tags": ["packages"],
-  "tasks": {}
-}
