+++ packages/app-pinecone/package.json
@@ -17,7 +17,7 @@
   },
   "dependencies": {
     "@pinecone-database/pinecone": "^6.1.3",
-    "@repo/app-config": "workspace:*",
+    "@repo/app-validation": "workspace:*",
     "@vendor/pinecone": "workspace:*"
   },
   "devDependencies": {
